requirements.txt
absl-py==1.4.0
addict==2.4.0
aenum==3.1.12
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alt-profanity-check==1.2.2
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.10.1
argilla==1.7.0
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backoff==2.2.1
basicsr==1.4.2
bcrypt==3.2.0
beautifulsoup4==4.12.2
better-profanity==0.6.1
bioc==2.0
bitsandbytes==0.39.0
blendmodes==2022
blessed==1.20.0
blinker==1.4
boltons==23.0.0
boto3==1.26.101
botocore==1.29.101
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
clean-fid==0.1.35
click==8.1.3
clickhouse-connect==0.5.25
cloud-init==23.1.2
cmake==3.26.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
ctranslate2==3.14.0
cycler==0.11.0
dataclasses-json==0.5.7
datasets==2.12.0
dbus-python==1.2.18
decorator==4.4.2
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.6
distro==1.7.0
distro-info===1.1build1
docutils==0.19
duckdb==0.8.0
effdet==0.4.1
einops==0.6.1
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.1.1
execnet==1.9.0
facexlib==0.3.0
faiss-gpu==1.7.2
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
filterpy==1.4.5
fire==0.5.0
flatbuffers==23.3.3
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
gradio_client==0.2.5
greenlet==2.0.2
grpcio==1.54.0
grpcio-tools==1.51.1
h11==0.14.0
h2o-wave @ file:///home/piratos/projects/h2o-llmstudio/deps/h2o_wave-nightly-py3-none-manylinux1_x86_64.whl
hf-hub-ctranslate2==2.0.9
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.20.2
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.15.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.28.0
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
iniconfig==2.0.0
inquirer==3.1.3
intervaltree==3.1.0
invisible-watermark==0.1.5
iopath==0.1.10
iotop==0.6
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.2.0
jq==1.4.1
jsonlines==3.1.0
jsonmerge==1.9.0
jsonpatch==1.32
jsonpointer==2.0
jsonref==1.1.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
kornia==0.6.12
langchain==0.0.178
lark==1.1.5
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
linkify-it-py==2.0.0
lit==16.0.2
llvmlite==0.39.1
lm-dataformat==0.0.20
lmdb==1.4.1
loralib==0.1.1
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
mypy-protobuf==3.4.0
neptune==1.1.1
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numba==0.56.4
numexpr==2.8.4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
onnx==1.13.1
onnxruntime==1.14.1
openai==0.27.6
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.2
orjson==3.8.11
packaging==23.1
pandoc==2.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
peft @ git+https://github.com/huggingface/peft.git@3714aa2fff158fdfa637b2b65952580801d890b2
pexpect==4.8.0
piexif==1.1.3
Pillow==9.5.0
pip-licenses==4.3.0
pluggy==1.0.0
plumbum==1.8.1
ply==3.11
portalocker==2.7.0
prettytable==3.7.0
protobuf==4.23.0
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.15.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pynvml==11.5.0
pyOpenSSL==23.1.1
pypandoc==1.11
pypandoc-binary==1.11
pyparsing==2.4.7
pypdf==3.8.1
pyphen==0.14.0
pyrsistent==0.18.1
pyserial==3.5
pytesseract==0.3.10
pytest==7.2.2
pytest-xdist==3.2.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-dotenv==1.0.0
python-editor==1.0.4
python-linux-procfs==0.6.3
python-magic==0.4.24
python-multipart==0.0.6
python-pptx==0.6.21
pytorch-lightning==1.7.7
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyudev==0.22.0
PyWavelets==1.4.1
PyYAML==6.0
readchar==4.0.5
realesrgan==0.3.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
resize-right==0.0.2
responses==0.18.0
rfc3986==1.5.0
rfc3987==1.3.8
rich==13.0.1
rouge-score==0.1.2
rsa==4.9
s3transfer==0.6.0
sacrebleu==2.3.1
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
service-identity==18.1.0
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.4
soupsieve==2.4.1
SQLAlchemy==2.0.15
ssh-import-id==5.11
starlette==0.26.1
strict-rfc3339==0.7
swagger-spec-validator==3.0.3
sympy==1.11.1
systemd-python==234
tabulate==0.9.0
tb-nightly==2.13.0a20230430
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
textstat==0.7.3
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
tqdm==4.65.0
trampoline==0.1.2
transformers==4.29.2
triton==2.0.0
Twisted==22.1.0
typer==0.9.0
types-protobuf==4.23.0.1
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.1
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
unstructured==0.6.6
unstructured-inference==0.4.4
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
wadllib==1.3.6
Wand==0.6.11
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.5.1
websockets==11.0.2
wrapt==1.14.1
xkit==0.0.0
XlsxWriter==3.1.1
xxhash==3.2.0
yapf==0.33.0
yarl==1.9.2
zipp==1.0.0
zope.interface==5.4.0
zstandard==0.21.0
